ml/pyproject.toml
[tool.poetry]
name = "tuo_ml"
version = "0.0.0"
description = "Machine Learning for TUO"
authors = ["Alexander Puck Neuwirth <alexander@neuwirth-informatik.de>"]
readme = "README.md"
repository = "https://github.com/APN-Pucky/tyrant_optimizer"

[tool.poetry.dependencies]
python = "^3.8"
deprecation = "*"

pyyaml="*"
tqdm="*"
numpy="^1.24"
scikit-learn="^1.1"
xgboost="^1.6"
sklearn2pmml="^0.92"
pandas="^1.5"
smpl_io="^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.masonry.api"